% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ref_groups.R
\name{create_ref_groups}
\alias{create_ref_groups}
\title{Create the refinement groups}
\usage{
create_ref_groups(emdata, config)
}
\description{
This function simulates the MARE2DEM's
refinement groups creation
}
