% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_unnest.R
\name{group_nest_dt}
\alias{group_nest_dt}
\title{Nest a data.table by group}
\usage{
group_nest_dt(dt, ..., .key = "data")
}
\description{
This function was originally written by TS Barrett in a document available at https://osf.io/f6pxw/download.
}
