% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_unnest.R
\name{unnest_dt}
\alias{unnest_dt}
\title{Unnest a data.table}
\usage{
unnest_dt(dt, col, id)
}
\description{
This function was originally written by TS Barrett in a document available at https://osf.io/f6pxw/download.
}
